refactor fred provider directory
