+++ openbb_sdk/providers/fred/openbb_fred/models/cpi.py
