+++ openbb_sdk/providers/fred/openbb_fred/provider.py
@@ -5,7 +5,7 @@
 # IMPORT INTERNAL
 from openbb_provider.abstract.provider import Provider, ProviderNameType
 
-from openbb_fred.cpi import FREDCPIFetcher
+from openbb_fred.models.cpi import FREDCPIFetcher
 
 # mypy: disable-error-code="list-item"
