+++ openbb_sdk/providers/fred/openbb_fred/utils/fred_helpers.py
@@ -1,22 +1,9 @@
 # IMPORT STANDARD
 import csv
-import pathlib
 from typing import List, Literal
 
 # IMPORT THIRD-PARTY
 # IMPORT INTERNAL
-
-
-BASE_DIR = (
-    pathlib.Path(__file__).parent.parent.parent.parent.parent
-    / "openbb_terminal"
-    / "economy"
-    / "datasets"
-)
-
-harmonized_cpi_path = BASE_DIR / "harmonized_cpi.csv"
-cpi_path = BASE_DIR / "cpi.csv"
-
 
 CPI_COUNTRIES = Literal[
     "australia",
@@ -78,8 +65,8 @@ CPI_FREQUENCY = Literal["monthly", "quarterly", "annual"]
 def all_cpi_options(harmonized: bool = False) -> List[dict]:
     data = []
 
-    the_path = harmonized_cpi_path if harmonized else cpi_path
-    with open(the_path, encoding="utf-8") as csv_file_handler:
+    file = "harmonized_cpi.csv" if harmonized else "cpi.csv"
+    with open(file, encoding="utf-8") as csv_file_handler:
         csv_reader = csv.DictReader(csv_file_handler)
         for rows in csv_reader:
             row = {key.lstrip("\ufeff"): value for key, value in rows.items()}
